Add logging of __int64 instead of ssize_t

+++ src/log.h
@@ -89,7 +89,7 @@ public:
 	log_record_stream& operator<<(double val);
 	log_record_stream& operator<<(bool val);
 #ifdef WIN32
-	log_record_stream& operator<<(ssize_t val);
+	log_record_stream& operator<<(__int64 val);
 #endif
 	log_record_stream& operator<<(std::streampos val);
 	log_record_stream& operator<<(const std::string& str);
